Don't redirect to /home when visiting /workspaces
