+++ typescript/web/src/components/home/home.tsx
@@ -1,13 +1,11 @@
-import { useBoolean } from "@chakra-ui/react";
 import { isEmpty, isNil } from "lodash/fp";
 import { useRouter } from "next/router";
 import { useEffect } from "react";
 import { useLastWorkspaceId, useOptionalWorkspaces } from "../../hooks";
 import { LayoutSpinner } from "../spinner";
-import { CreateWorkspaceModal } from "../workspace-switcher/create-workspace-modal";
 import { Workspaces } from "../workspaces";
 
-const useLastWorkspaceUrl = (): string | undefined => {
+const useLastWorkspaceIdUrl = (): string | undefined => {
   const workspaces = useOptionalWorkspaces();
   const lastWorkspaceId = useLastWorkspaceId();
   if (isNil(workspaces) || isEmpty(lastWorkspaceId)) return undefined;
@@ -24,21 +22,7 @@ const useRedirectTo = (redirectTo: string | undefined) => {
 };
 
 export const Home = () => {
-  const redirectTo = useLastWorkspaceUrl();
+  const redirectTo = useLastWorkspaceIdUrl();
   useRedirectTo(redirectTo);
-  const [
-    showCreateWorkspaceModal,
-    { on: openCreateWorkspaceModal, off: closeCreateWorkspaceModal },
-  ] = useBoolean(false);
-  const workspaces = useOptionalWorkspaces();
-  if (isNil(workspaces) || !isNil(redirectTo)) return <LayoutSpinner />;
-  return (
-    <>
-      <Workspaces openCreateWorkspaceModal={openCreateWorkspaceModal} />
-      <CreateWorkspaceModal
-        isOpen={showCreateWorkspaceModal}
-        onClose={closeCreateWorkspaceModal}
-      />
-    </>
-  );
+  return <>{isNil(redirectTo) ? <Workspaces /> : <LayoutSpinner />}</>;
 };
