+++ typescript/web/src/pages/workspaces/[...rest].tsx
